requirements.txt
Flask==3.1.2
qrcode[pil]==8.2
Pillow==12.1.0
gunicorn
